e000-01-r3gan-baseline/requirements.txt
torch>=2.1.0
torchvision>=0.16.0
torchmetrics[image]>=1.4.0
torch-fidelity>=0.3.0
pyyaml>=6.0

